fix bug

+++ lib/model/response/coins_current_data_response.dart
@@ -70,7 +70,7 @@ class CoinsCurrentDataResponse {
     double? developerScore;
     double? communityScore;
     double? liquidityScore;
-    int? publicInterestScore;
+    double? publicInterestScore;
     MarketData? marketData;
     CommunityData? communityData;
     DeveloperData? developerData;
